DCPlatform/requirements.txt
alembic==1.4.2
click==7.1.1
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==0.14.1
Mako==1.1.2
MarkupSafe==1.1.1
numpy==1.18.2
pandas==1.0.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.16
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.2.1
nltk
scikit-learn
python-Levenshtein
numpy
scipy
pandas
langdetect


greenlet==1.1.0
transformers[torch]
scipy
sentence-transformers
openml
pymfe
protobuf
